GMLAS: compiler warning fixes
git-svn-id: https://svn.osgeo.org/gdal/trunk@35803 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/gmlas/ogr_gmlas.h
@@ -843,7 +843,7 @@ class GMLASSchemaAnalyzer
 
         CPLString AddSerialNumber(const CPLString& osNameIn,
                                   int iOccurrence,
-                                  int nOccurrences);
+                                  size_t nOccurrences);
 
         void CollectClassesReferences(
                                 GMLASFeatureClass& oClass,
